pyproject.toml
[project]
name = "pptxai1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mss>=9.0.1",
    "python-pptx>=1.0.2",
    "python-dotenv>=1.0.1",
    "pydantic>=2.8.0",
    "openai>=1.44.0",
    "pillow>=10.4.0",
    "pytest>=9.0.1",
    "pdf2image>=1.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
pythonpath = ["."]
